fix(ts): align custom-field filter/sort typings with current master
